Move serve functionality out of allennlp command.
The `allennlp` command currently requires dependencies that few of our
users will use and that increase the time it takes to run `allennlp`.
Furthermore, the serve command does not work from pip installs, which is
the main way our users install AllenNLP.
