+++ allennlp/service/predictors/__init__.py
@@ -8,7 +8,7 @@ a ``Predictor`` that wraps it.
 """
 import warnings
 
-from allennlp.predictors.predictor import Predictor, DemoModel
+from allennlp.predictors.predictor import Predictor
 from allennlp.predictors.bidaf import BidafPredictor
 from allennlp.predictors.constituency_parser import ConstituencyParserPredictor
 from allennlp.predictors.coref import CorefPredictor
