+++ allennlp/service/predictors/predictor.py
@@ -1,6 +1,6 @@
 # pylint: disable=unused-import
 import warnings
 
-from allennlp.predictors.predictor import Predictor, DemoModel, DEFAULT_PREDICTORS
+from allennlp.predictors.predictor import Predictor, DEFAULT_PREDICTORS
 warnings.warn("allennlp.service.predictors.* has been deprecated."
               " Please use allennlp.predictors.*", FutureWarning)
