+++ allennlp/tests/models/sniff_test.py
@@ -1,6 +1,6 @@
 # pylint: disable=no-self-use,line-too-long
 
-from allennlp.commands.serve import DEFAULT_MODELS
+from allennlp.service.server_flask import DEFAULT_MODELS
 from allennlp.common.testing import AllenNlpTestCase
